+++ pkg/bigquery/macros.go
@@ -21,6 +21,14 @@ func macroTimeGroup(query *sqlds.Query, args []string) (string, error) {
 	if len(args) < 2 {
 		return "", fmt.Errorf("%w: expected 2 arguments, received %d", errors.New("macro $__timeGroup needs time column and interval"), len(args))
 	}
+	
+	if(args[0] == "") {
+		return "", fmt.Errorf("the first paramater(time column) for $__timeGroup macro cannot be empty")
+	}
+
+	if (args[1] == "") {
+		return "", fmt.Errorf("the second paramater(interval) for $__timeGroup macro cannot be empty")
+	}
 
 	timeVar := args[0]
 	intervalVar := strings.Trim(args[1], "'\"")